once again.

+++ DPOC/src/Attack.cpp
@@ -85,7 +85,7 @@ int calculate_physical_damage(Character* attacker, Character* target, Item* weap
     damage = ((b / 2.0f + 1.0f) * (float)random_range(0, 256) / 256.0f + 2.0f) / 3.0f;
   }
 
-  if (damage <= 0)
+  if ((int)damage <= 0)
   {
     damage = random_range(0, 2);
   }
